Move misplaced code
