+++ src/coreclr/jit/jitpch.h
@@ -16,4 +16,3 @@
 
 #include "jitconfig.h"
 #include "compiler.h"
-#include "phase.h"
